Convert to `pyproject.toml` project, and drop Python <3.8
Nothing should prevent this plugin from continuing to work with Sopel
7.1, but I'm no longer testing it on Python 2. People should at least
upgrade to Python 3, even if they don't want Sopel 8 yet.
