+++ NEWS
@@ -1,5 +1,6 @@
-sopel-flipper 0.1.0
-===================
+## Changelog
+
+### 0.1.0
 
 Initial release.
