+++ pyproject.toml
@@ -1,0 +1,53 @@
+[build-system]
+requires = ["setuptools>=63.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools]
+platforms = ["Linux x86, x86-64"]
+
+[tool.setuptools.packages.find]
+include = ["sopel_flipper", "sopel_flipper.*"]
+namespaces = false
+
+[tool.setuptools.dynamic]
+readme = { file=["README.md", "NEWS"], content-type="text/markdown" }
+
+[project]
+name = "sopel-flipper"
+version = "1.0.0.dev0"
+description = "A Sopel plugin that flips text in response to CTCP ACTIONs."
+
+authors = [
+  { name="dgw", email="dgw@technobabbl.es" },
+]
+
+license = { text="EFL-2.0" }
+dynamic = ["readme"]
+
+classifiers = [
+  "Intended Audience :: Developers",
+  "Intended Audience :: System Administrators",
+  "License :: Eiffel Forum License (EFL)",
+  "License :: OSI Approved :: Eiffel Forum License",
+  "Programming Language :: Python :: 3 :: Only",
+  "Topic :: Communications :: Chat :: Internet Relay Chat",
+]
+keywords = [
+  "sopel",
+  "plugin",
+  "bot",
+  "irc",
+]
+
+requires-python = ">=3.8, <4"
+dependencies = [
+  "sopel>=7.1",
+  "upsidedown==0.4",
+]
+
+[project.urls]
+"Homepage" = "https://github.com/dgw/sopel-flipper"
+"Bug Tracker" = "https://github.com/dgw/sopel-flipper/issues"
+
+[project.entry-points."sopel.plugins"]
+"flipper" = "sopel_flipper.plugin"
