--- setup.cfg
@@ -1,26 +1,0 @@
-[metadata]
-name = sopel-flipper
-version = 0.1.0
-description = A Sopel plugin that flips text in response to CTCP ACTIONs.
-author = dgw
-author_email = dgw@technobabbl.es
-url = https://github.com/dgw/sopel-flipper
-license = Eiffel Forum License, version 2
-classifiers =
-    Intended Audience :: Developers
-    Intended Audience :: System Administrators
-    License :: Eiffel Forum License (EFL)
-    License :: OSI Approved :: Eiffel Forum License
-    Topic :: Communications :: Chat :: Internet Relay Chat
-
-[options]
-packages = find:
-zip_safe = false
-include_package_data = true
-install_requires =
-    sopel>=7.1,<9
-    upsidedown
-
-[options.entry_points]
-sopel.plugins =
-    flipper = sopel_flipper
